add minimal support for installation via pip
Currently mimic consists solely of one script which is marked as executable.
This is a non-idiomatic way to distribute Python command-line scripts. Add a
simple ``setup.py`` (which will need augmenting with author info, etc) and
re-factor mimic into a single-file module.

The upshot of this is that one can now install mimic directly from pip via

    pip install git+git://github.com/reinderien/mimic.git

and have setuptools create a mimic command which automatically invokes the
correct Python.

For development the usual

    pip install -e .

can be used to use symlink magic to make sure that changes to
``magic/__init__.py`` are respected when launching the mimic wrapper script.

Update README.md with the installation incantation.

A ``__main__.py`` file is also included which allows execution of mimic via

    python -m mimic

One could also add a wrapper script to the repository for this but that's
setuptools' job really.