add docutils as a requirement
Issue #4 notes that docutils is a requirement. Reflect this in the ``setup.py``
script. For the moment leave the same information in ``requirements.txt``. The
difference between the two approaches is subtle and, unless one is sure, it is
probably worth the hit of being a little wet (i.e. non-DRY).